requirements.txt
accelerate>=0.26.1
aiohttp
auto-gptq
bcembedding
beautifulsoup4
duckduckgo_search
einops
faiss-gpu
langchain>=0.1.12
loguru
lxml_html_clean
openai>=1.0.0
openpyxl
pandas
pydantic>=1.10.13
pymupdf
python-docx
pytoml
readability-lxml
redis
requests
scikit-learn
sentence_transformers==2.2.2
# See https://github.com/deanmalmgren/textract/issues/461
# textract @ git+https://github.com/tpoisonooo/textract@master
textract
tiktoken
torch>=2.0.0
transformers>=4.37.0
transformers_stream_generator
unstructured